+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.5.21
+
+* `Versions`
+    * `Klock`: `2.3.1` -> `2.3.2`
+* `Serialization`
+    * `Typed Serializer`:
+        * `TypedSerializer` Descriptor serial name has been fixed
+
 ## 0.5.20
 
 * `Repos`: